refactored HighEntropyStringsFinder
